DimMat

Description

see examples/

Quasiquotes

matD :: QuasiQuoter

blockD :: QuasiQuoter

Data.Packed.Vector

(@>) :: (HNat2Integral i, AtEq r i ri, MultEq r1 ri u) => D `(r1, `[r]`)` a -> Proxy i -> Quantity u a

Data.Packed.Vector.@>

Data.Packed.Matrix

dimension

cols :: D t a -> Int

rows :: D t a -> Int

colsNT :: D `(a, r : (c : cs))` e -> Proxy (HLength (DOne : c))

H.cols except type-level

rowsNT :: D `(a, r : c)` e -> Proxy (HLength (a : ri))

H.rows except type-level

hasRows :: (SameLength (HReplicateR n ()) r, HLength r ~ n) => D `(a, ra : ca)` e -> Proxy (n :: HNat) -> D `(a, ra : ca)` e

(m hasRows n) constrains the matrix/vector m to have n rows

hasCols :: (SameLength (HReplicateR n ()) ci, HLength ci ~ n) => D `(a, ra : (ca : rest))` e -> Proxy (n :: HNat) -> D `(a, ra : (ca : rest))` e

(m hasRows n) constrains the matrix/vector m to have n rows

trans :: Trans a b => D a e -> D b e

class ToHLists sh e xs where

Methods

toHLists :: D sh e -> HList xs

toHList :: ToHList sh e result => D sh e -> HList result

class FromHLists lists sh e where

Methods

fromHLists :: HList lists -> D sh e

fromHList :: FromHList list sh e => HList list -> D sh e

(@@>) :: (HNat2Integral i, HNat2Integral j, AtEq2 (r1 : r) i (DOne : c) j ty) => D `(r1, `[r, c]`)` a -> (Proxy i, Proxy j) -> Quantity ty a

Data.Packed.Matrix.@@>

diagBlock :: (HMapOut UnDimMat (b : bs) (Matrix e), Num e, Field e, AppendShOf b bs (D `(a, sh)` e), sh ~ `[r, c]`) => HList (b : bs) -> D `(a, sh)` e

blockDiag. The blocks should be provided as:

blockDiag $ hBuild m1 m2 m3

only available if hmatrix >= 0.15

ident :: forall ones e. (Field e, HNat2Integral (HLength (DOne : ones))) => D `(DOne, `[ones, ones]`)` e

identity matrix. The size is determined by the type.

Numeric.Container

diag :: (MapConst DOne r ~ c, SameLength r c) => D `(a, `[r]`)` t -> D `(a, `[r, c]`)` t

ctrans :: D ((,) * [[*]] r1 ((:) [*] x ((:) [*] y ([] [*])))) a -> D ((,) * [[*]] r1 ((:) [*] y ((:) [*] x ([] [*])))) a

conjugate transpose

Container class

scalar :: Scalar as => D `(a, [])` e -> D `(a, as)` e

conj :: D sh a -> D sh a

scale :: Scale a b c => Quantity a e -> D b e -> D c e

scaleRecip :: ScaleRecip a b c => D `(a, [])` e -> D b e -> D c e

recipMat :: forall b c e. (Field e, ScaleRecip DOne b c) => D b e -> D c e

a shortcut for scaleRecip (DScal 1)

addConstant :: (Field e, CanAddConst a sh) => D `(a, [])` e -> D `(a, sh)` e -> D `(a, sh)` e

add :: D a e -> D a e -> D a e

sub :: D a e -> D a e -> D a e

mulMat :: (MultEq as bs cs, MultEq a b c, cs ~ `[t1, t2]`) => D `(a, as)` e -> D `(b, bs)` e -> D `(c, cs)` e

mulVec :: (MultEq as bs cs, MultEq a b c, cs ~ `[t1]`) => D `(a, as)` e -> D `(b, bs)` e -> D `(c, cs)` e

divideMat :: (MultEq as cs bs, MultEq a c b, cs ~ `[t1, t2]`) => D `(a, as)` e -> D `(b, bs)` e -> D `(c, cs)` e

divideVec :: (MultEq as cs bs, MultEq a c b, cs ~ `[t1]`) => D `(a, as)` e -> D `(b, bs)` e -> D `(c, cs)` e

equal :: D a e -> D a e -> Bool

arctan2 :: (bs ~ MapMapConst DOne as) => D `(a, as)` e -> D `(a, as)` e -> D `(b, bs)` e

hconcat :: (MultEq (rem :: *) a b, MultEq rem ra rb, MultEq rem (DOne : cb) cb', AppendEq ca cb' cc) => D `(a, `[ra, ca]`)` e -> D `(b, `[rb, cb]`)` e -> D `(a, `[ra, cc]`)` e

the slightly involved type here exists because ci1 and ci2 both start with DOne, but ci2's contribution to ci3 does not need a DOne at the start. Another way to read the constraints here is:

map (*rem) (a11 : ri) = b11 : bi
ci3 = ci1 ++ map (*rem) ci2

The same idea happens with vconcat.

vconcat :: AppendEq ra (b : rb) rc => D `(a, `[ra, ca]`)` e -> D `(b, `[rb, ca]`)` e -> D `(a, `[rc, ca]`)` e

cmap :: CMap f a b e e' => f -> D a e -> D b e'

konst :: forall e a ra ca. (Field e, HNat2Integral (HLength (a : ra)), HNat2Integral (HLength (DOne : ca)), AllEq DOne ca, AllEq a ra) => D `(a, [])` e -> D `(a, `[ra, ca]`)` e

Numeric.Container.konst, but the size is determined by the type.

>>> let n = hSucc (hSucc hZero) -- 2
>>> konst ((1::Double) *~ second) `hasRows` n `hasCols` n
2><2 1   1  
s    1.0 1.0
s    1.0 1.0

zeroes :: forall c a r e. (Field e, HNat2Integral (HLength (a : r)), HNat2Integral (HLength (DOne : c))) => D `(a, `[r, c]`)` e

zero matrix. The size and dimension is determined by the type.

Product class

dot :: (Dot a b c, Element e) => D a e -> D b e -> D c e

pnorm :: (AllEq r1 r, AllEq DOne c) => NormType -> D `(r1, `[r, c]`)` a -> Quantity r1 (RealOf a)

Random numbers

Element conversion

Input/Output

Experimental

Numeric.LinearAlgebra.Algorithms

incomplete wrapper for Numeric.LinearAlgebra.Algorithms

Linear Systems

inv :: (PInv a b, b ~ `(t1, `[t2, t3]`)`) => D a e -> D b e

class PInv a b where

type for a pseudoinverse (and inverse):

The single instance comes from looking at inverses from a 2x2 matrix (let's call A):

a b
c d

and the inverse * determinant of the original

 d  -b
-c   a

In the product A * A^-1 the diagonal is dimensionless (DOne).

That happens if the row and column type-level unit lists are reciprocals of eachother (AreRecips), so the constraint on the instance of PInv encodes this exactly (plus some constraints requiring that sh and sh' are at least 1x1)

Methods

pinv :: D a e -> D b e

pinvTol :: (PInv a b, b ~ `(t1, `[t2, t3]`)`) => Double -> D a e -> D b e

det :: Det a b => D a e -> Quantity b e

rank :: D t a -> Int

Matrix factorizations

Singular value decomposition

Eigensystems

QR

Cholesky

Hessenberg

Schur

LU

Matrix functions

expm :: AreRecips r c => D `(r1, `[r, c]`)` a -> D `(r1, `[r, c]`)` a

Numeric.LinearAlgebra.Algorithms.expm

y t = expm (scale t a) `multiply` y0 solves the DE y' = Ay where y0 is the value of y at time 0

Nullspace

Norms

Misc

Util

Automatic Differentiation

to keep types looking ok

module Data.HList.CommonMain

data Complex a :: * -> *

Complex numbers are an algebraic type.

For a complex number z, abs z is a number with the magnitude of z, but oriented in the positive real direction, whereas signum z has the phase of z, but unit magnitude.