DimMat.Internal

Description

an incomplete extension of dimensional-tf to work with hmatrix and ad Haddocks are organized to follow hmatrix

note: for subscripting, use the HNat while the units still use NumType

TODO:

Friendly syntax for introduction (more matD)

Friendly syntax for elimination (matD as pattern)

A pretty-printer that multiplies out the dimensions at each place? The current show instance makes you mentally multiply out row and column units (which helps you to see the big picture, but may be more work in other cases)

check that all types that could could be inferred are

default columns/rows to dimensionless?

missing operations (check export list comments)

Replacements for Dimensional classes

class Mul a b c => Mul a b c

a version of Numeric.Units.Dimensional.Mul which requires the arguments to include the Dim type constructor

class Div a b c => Div a b c

a version of Numeric.Units.Dimensional.Div which requires the arguments to include the Dim type constructor

AD

TODO: gradients, hessians, etc.

Types for derivative towers can see hlist's HList/Data/HList/broken/Lazy.hs, but laziness doesn't really make much sense if the take that is eventually used to get a finite list for printing etc.

Complications include the fact that AD.grad needs a traversable, but hmatrix stuff is not traversable (due needing Storable). In ipopt-hs I got around this problem by copying data. Perhaps that is the solution?

BROKEN
grad :: (Num a, AreRecips i iinv, H.Element a, Storable a,
          MapMultEq o iinv r) =>
        (forall s. (AD.Mode s, H.Container H.Vector (AD.AD s a),
                    Storable (AD.AD s a), H.Field (AD.AD s a))
                => DimMat '[i] (AD.AD s a)
                -> Quantity o (AD.AD s a))
     -> DimMat '[i] a
     -> DimMat '[r] a
grad f (DimVec x) = DimMat (H.fromLists [AD.grad (unQty . f . DimVec . H.fromList) (H.toList x)])
    where unQty (Dimensional a) = a

GADT for linear algebra with units

data D sh e where

Generalization of Dimensional to matrices and vectors. Units in each coordinate are known at compile-time. This wraps up HMatrix.

DMat

the units at coordinate ij are (r1 ': r)_i (DOne ': c)_j

DVec

the units at coordinate i are (r1 ': r)_i

DScal

is the same as Dimensional

Constructors

DMat :: (Container Matrix e, Field e) => Matrix e -> D `(r1, `[r, c]`)` e

DVec :: (Container Vector e, Field e) => Vector e -> D `(r1, `[r]`)` e

DScal :: Field e => e -> D `(r1, [])` e

pretty instance

pad :: [String] -> [String]

class PPUnits sh where

Methods

ppUnits :: Proxy sh -> [[String]]

class PPUnits' sh where

Methods

ppUnits' :: Proxy sh -> [String]

class ShowDimSpec a where

Methods

showDimSpec :: Proxy a -> String

Constraints

A major theme in this library is that type inference goes in whichever direction it can: in ordinary haskell it is very common for argument types to be determined by the result types. For example see any code that uses Num or Read.

When we use type families, things look more convenient:

data Nat = S Nat | Z

type family Add (a :: Nat) (b :: Nat) :: Nat
type instance Add Z b = b
type instance Add (S a) b = Add a (S b)

But ghc is unable to deduce things like a ~ Z given evidence such as Add Z a ~ Z. One way around this is to use ConstraintKinds:

type AddT (a :: Nat) (b :: Nat) (c :: Nat)
          = (Add a b ~ c, Sub c a ~ b, Sub c b ~ a)

Which leads to functions like f :: AddT a b ab => ... . This is bad for a couple reasons:

the right-hand-side of the type can only mention type variables on the left-hand-side.

the left hand side can only bind type variables. Working around this leads to many auxiliary type families such as Fst, Snd and Head, or leads to a type family AddT

So below many constraints expressed as classes, since they have less of those limitations.

class MultEq a b c

a*b = c when any are lists

Zip

class (SameLength a b, SameLength b c) => Zip3 op a b c

class SameLength a b => Zip2 op a b c

class Zip1 op a b c

class SameLength a ab => Outer a b ab

given ijs :: [[Quantity a]] (except the : and [] constructors are actually (,) and (), ie. a HList that doesn't use the HList constructors), calculate a DimMat rowUnits colUnits, where the outer product of rowUnits and colUnits gives the units at each index in the ijs. The first element of colUnits is DOne.

DimMatFromTuple

class DimMatFromTuple ijs r1 r c e

type family TupleToHListU a :: [*]

type family TuplesToHListU a :: [[*]]

type family DMFromTuple1 e b b' :: Constraint

helper for DimMatFromTuple

type family DMFromTuple2 e b b' :: Constraint

type family DMFromTuple3 e b b' :: Constraint

toDM :: DimMatFromTuple ijs r1 r c e => ijs -> Proxy (D `(r1, `[r, c]`)` e)

just for types produced by the matD quasiquote

type family InnerCxt t a b :: Constraint

InnerCxt t a b = t ~ dot a b

class (SameLength a b, InnerCxt c a b) => Inner a b c

class ProdEq a b

ProdEq a b is product a ~ b

class RecipEq a aInv

RecipEq a aInv is a*aInv ~ DOne (or a list of DOne)

type family AtEq2 a n b m c :: Constraint

AtEq a n b m c calculates (At a n `Mult` At b m) ~ c, but also can infer part of the a if the b and c are known

type family AtEq a n b :: Constraint

(@>) :: (HNat2Integral i, AtEq r i ri, MultEq r1 ri u) => D `(r1, `[r]`)` a -> Proxy i -> Quantity u a

Data.Packed.Vector.@>

(@@>) :: (HNat2Integral i, HNat2Integral j, AtEq2 (r1 : r) i (DOne : c) j ty) => D `(r1, `[r, c]`)` a -> (Proxy i, Proxy j) -> Quantity ty a

Data.Packed.Matrix.@@>

pnorm :: (AllEq r1 r, AllEq DOne c) => NormType -> D `(r1, `[r, c]`)` a -> Quantity r1 (RealOf a)

class AllEq a xs

AllEq a xs is like all (all (a==)) xs, all (a ==) xs, a == xs: whichever amount of [ ] is peeled off before making the comparison (with ~)

class Dot a b c where

c = a dot b is one of:

c_ij = sum_j a_ij b_jk
c_k  = sum_j a_j  b_jk
c_i  = sum_j a_ij b_j
c    = sum_j a_j  b_j

Methods

dot :: Element e => D a e -> D b e -> D c e

class Trans a b where

Methods

trans :: D a e -> D b e

class PInv a b where

type for a pseudoinverse (and inverse):

The single instance comes from looking at inverses from a 2x2 matrix (let's call A):

a b
c d

and the inverse * determinant of the original

 d  -b
-c   a

In the product A * A^-1 the diagonal is dimensionless (DOne).

That happens if the row and column type-level unit lists are reciprocals of eachother (AreRecips), so the constraint on the instance of PInv encodes this exactly (plus some constraints requiring that sh and sh' are at least 1x1)

Methods

pinv :: D a e -> D b e

type family LengthSndTwo a :: Constraint

type AreRecips a b = MultEq a b DOne

inv :: (PInv a b, b ~ `(t1, `[t2, t3]`)`) => D a e -> D b e

pinvTol :: (PInv a b, b ~ `(t1, `[t2, t3]`)`) => Double -> D a e -> D b e

class Det a b where

Methods

det :: D a e -> Quantity b e

expm :: AreRecips r c => D `(r1, `[r, c]`)` a -> D `(r1, `[r, c]`)` a

Numeric.LinearAlgebra.Algorithms.expm

y t = expm (scale t a) `multiply` y0 solves the DE y' = Ay where y0 is the value of y at time 0

class Scale a b c where

Numeric.Container.scale

Methods

scale :: Quantity a e -> D b e -> D c e

fromQty :: Field e => Quantity a e -> D `(a, [])` e

toQty :: D `(a, [])` e -> Quantity a e

class ScaleRecip a b c where

Numeric.Container.scaleRecip

Methods

scaleRecip :: D `(a, [])` e -> D b e -> D c e

class ScaleRecip1 bool a b c where

Methods

scaleRecip1 :: Proxy bool -> D `(a, [])` e -> D b e -> D c e

type ScaleRecipCxt r1 r1' r r' rs rs' a b c = (MultEq a (r1' : r') (r1 : r), MultEq rs rs' DOne, b ~ `(r1, r : rs)`, c ~ `(r1', r' : rs')`)

recipMat :: forall b c e. (Field e, ScaleRecip DOne b c) => D b e -> D c e

a shortcut for scaleRecip (DScal 1)

liftH2 :: (forall h f. (Container f e, h ~ f e) => h -> h -> h) -> D a e -> D a e -> D a e

add :: D a e -> D a e -> D a e

sub :: D a e -> D a e -> D a e

mulMat :: (MultEq as bs cs, MultEq a b c, cs ~ `[t1, t2]`) => D `(a, as)` e -> D `(b, bs)` e -> D `(c, cs)` e

mulVec :: (MultEq as bs cs, MultEq a b c, cs ~ `[t1]`) => D `(a, as)` e -> D `(b, bs)` e -> D `(c, cs)` e

divideMat :: (MultEq as cs bs, MultEq a c b, cs ~ `[t1, t2]`) => D `(a, as)` e -> D `(b, bs)` e -> D `(c, cs)` e

divideVec :: (MultEq as cs bs, MultEq a c b, cs ~ `[t1]`) => D `(a, as)` e -> D `(b, bs)` e -> D `(c, cs)` e

arctan2 :: (bs ~ MapMapConst DOne as) => D `(a, as)` e -> D `(a, as)` e -> D `(b, bs)` e

equal :: D a e -> D a e -> Bool

class CMap f a b e e' where

cmap f m gives a matrix m'

f is applied to

Methods

cmap :: f -> D a e -> D b e'

type family AppendEq' a b ab :: Constraint

type AppendEq a b ab = (ab ~ HAppendR a b, AppendEq' a b ab, SameLength (DropPrefix a ab) b, SameLength (DropPrefix b ab) a)

a bit overkill? a ++ b = ab

type family DropPrefix a ab :: [k]

hconcat :: (MultEq (rem :: *) a b, MultEq rem ra rb, MultEq rem (DOne : cb) cb', AppendEq ca cb' cc) => D `(a, `[ra, ca]`)` e -> D `(b, `[rb, cb]`)` e -> D `(a, `[ra, cc]`)` e

the slightly involved type here exists because ci1 and ci2 both start with DOne, but ci2's contribution to ci3 does not need a DOne at the start. Another way to read the constraints here is:

map (*rem) (a11 : ri) = b11 : bi
ci3 = ci1 ++ map (*rem) ci2

The same idea happens with vconcat.

vconcat :: AppendEq ra (b : rb) rc => D `(a, `[ra, ca]`)` e -> D `(b, `[rb, ca]`)` e -> D `(a, `[rc, ca]`)` e

rank :: D t a -> Int

cols :: D t a -> Int

rows :: D t a -> Int

rowsNT :: D `(a, r : c)` e -> Proxy (HLength (a : ri))

H.rows except type-level

colsNT :: D `(a, r : (c : cs))` e -> Proxy (HLength (DOne : c))

H.cols except type-level

hasRows :: (SameLength (HReplicateR n ()) r, HLength r ~ n) => D `(a, ra : ca)` e -> Proxy (n :: HNat) -> D `(a, ra : ca)` e

(m hasRows n) constrains the matrix/vector m to have n rows

hasCols :: (SameLength (HReplicateR n ()) ci, HLength ci ~ n) => D `(a, ra : (ca : rest))` e -> Proxy (n :: HNat) -> D `(a, ra : (ca : rest))` e

(m hasRows n) constrains the matrix/vector m to have n rows

class (MapConst [] as ~ as) => Scalar as where

H.scalar

Methods

scalar :: D `(a, [])` e -> D `(a, as)` e

konst :: forall e a ra ca. (Field e, HNat2Integral (HLength (a : ra)), HNat2Integral (HLength (DOne : ca)), AllEq DOne ca, AllEq a ra) => D `(a, [])` e -> D `(a, `[ra, ca]`)` e

Numeric.Container.konst, but the size is determined by the type.

>>> let n = hSucc (hSucc hZero) -- 2
>>> konst ((1::Double) *~ second) `hasRows` n `hasCols` n
2><2 1   1  
s    1.0 1.0
s    1.0 1.0

ident :: forall ones e. (Field e, HNat2Integral (HLength (DOne : ones))) => D `(DOne, `[ones, ones]`)` e

identity matrix. The size is determined by the type.

zeroes :: forall c a r e. (Field e, HNat2Integral (HLength (a : r)), HNat2Integral (HLength (DOne : c))) => D `(a, `[r, c]`)` e

zero matrix. The size and dimension is determined by the type.

type family CanAddConst a m :: Constraint

addConstant :: (Field e, CanAddConst a sh) => D `(a, [])` e -> D `(a, sh)` e -> D `(a, sh)` e

conj :: D sh a -> D sh a

ctrans :: D ((,) * [[*]] r1 ((:) [*] x ((:) [*] y ([] [*])))) a -> D ((,) * [[*]] r1 ((:) [*] y ((:) [*] x ([] [*])))) a

conjugate transpose

diag :: (MapConst DOne r ~ c, SameLength r c) => D `(a, `[r]`)` t -> D `(a, `[r, c]`)` t

diagBlock :: (HMapOut UnDimMat (b : bs) (Matrix e), Num e, Field e, AppendShOf b bs (D `(a, sh)` e), sh ~ `[r, c]`) => HList (b : bs) -> D `(a, sh)` e

blockDiag. The blocks should be provided as:

blockDiag $ hBuild m1 m2 m3

only available if hmatrix >= 0.15

data UnDimMat

Constructors

UnDimMat

class DiagBlock bs t

class AppendShOf a as aas

AppendShOf a [b,c,d] aas makes aas have the type of a matrix that has a,b,c,d along the diagonal

type family AppendDims a b c :: Constraint

class ToHList sh e result where

Methods

toHList :: D sh e -> HList result

class HListFromList e e' where

Methods

hListFromList :: [e] -> HList e'

type family ToHListRow a e b :: Constraint

data AddDimensional

Constructors

AddDimensional

class FromHList list sh e where

Methods

fromHList :: HList list -> D sh e

data RmDimensional

Constructors

RmDimensional

class FromHLists lists sh e where

Methods

fromHLists :: HList lists -> D sh e

newtype HMapOutWith f

Constructors

HMapOutWith f

class ToHListRows' r c e rows

data AddQty u

class ToHLists sh e xs where

Methods

toHLists :: D sh e -> HList xs

type family MapMapConst a xs :: [[k]]

\a xs -> map (map (const a)) xs

type family MapConst a xs :: [k]

\a xs -> map (const a) xs

type family FromPairs a :: [*]

convert from (a,(b,(c,(d,())))) to '[a,b,c,d]