DimMat

Description

see examples/

Documentation

module DimMat.Internal

module DimMat.QQ

module Numeric.Units.Dimensional.TF.Prelude

module Data.HList.CommonMain

to keep types looking ok

data Z

Type level zero.

data S n

Successor type for building type level natural numbers.

data N n

Negation type, used to represent negative numbers by negating type level naturals.

type Zero = Z

data Complex a

Complex numbers are an algebraic type.

For a complex number z, abs z is a number with the magnitude of z, but oriented in the positive real direction, whereas signum z has the phase of z, but unit magnitude.