DimMat.Internal

Description

an incomplete extension of dimensional-tf to work with hmatrix and ad Haddocks are organized to follow hmatrix

Friendly syntax for introduction (more matD)

Friendly syntax for elimination (matD as pattern)

A pretty-printer that multiplies out the dimensions at each place? The current show instance

A clean way to get the n x n dimensionless identity matrix

check that all types that could could be inferred are

default columns/rows to dimensionless?

missing operations (check export list comments)

Data.Packed.Vector

(@>) :: NumType i => DimMat `[units]` a -> i -> Quantity (At units i) a

Data.Packed.Matrix

dimension

cols :: DimMat t t1 -> Int

rows :: DimMat t t1 -> Int

colsNT :: DimMat `[ri, ci]` a -> Proxy (Len ci)

rowsNT :: DimMat `[ri, ci]` a -> Proxy (Len ri)

trans :: (Trans sh sh', sh' ~ `[_1 : __1, DOne : _2]`, sh ~ `[_3 : __3, DOne : _4]`) => DimMat sh a -> DimMat sh' a

(@@>) :: (NumType i, NumType j, AtEq ri i ci j ty) => DimMat `[ri, ci]` a -> (i, j) -> Quantity ty a

Data.Packed.Matrix.@@>

diagBlock :: (PairsToList t e, db ~ DimMat `[ri, DOne : ci]` e, Num e, Field e, DiagBlock t db) => t -> db

blockDiag. The blocks should be provided as:

 blockDiag (m1, (m2, (m3, ())))

XXX should we bring in HList for this?

only available if hmatrix >= 0.15

Numeric.Container

diag :: (MapConst DOne v ~ c, c ~ (DOne : _1)) => DimMat `[v]` t -> DimMat `[v, c]` t

ctrans :: (MapMultEq' * _3 _4 __1, MapMultEq' * _3 _2 __3, SameLength' [*] [*] _2 __3, SameLength' [*] [*] __3 _2, SameLength' [*] [*] __1 _4, SameLength' [*] [*] _4 __1) => DimMat (: [*] (: * _3 __3) (: [*] (: * DOne _4) ([] [*]))) a -> DimMat (: [*] (: * _3 __1) (: [*] (: * DOne _2) ([] [*]))) a

conjugate transpose

Container class

scalar :: (Field a, sh ~ `[`[u]`, `[DOne]`]`) => Quantity u a -> DimMat sh a

conj :: DimMat sh a -> DimMat sh a

scale :: MapMultEq e ri ri' => Quantity e a -> DimMat `[ri, ci]` a -> DimMat `[ri', ci]` a

Numeric.Container.scale

scaleRecip :: (MapMultEq e' ri ri', AreRecips e e') => Quantity e a -> DimMat `[ri, ci]` a -> DimMat `[ri', ci]` a

Numeric.Container.scaleRecip

addConstant :: (Field a, CanAddConst u sh) => Quantity u a -> DimMat sh a -> DimMat sh a

add :: Container Vector a => DimMat (: [*] ri (: [*] ci ([] [*]))) a -> DimMat (: [*] ri (: [*] ci ([] [*]))) a -> DimMat (: [*] ri (: [*] ci ([] [*]))) a

sub :: Container Vector a => DimMat (: [*] ri (: [*] ci ([] [*]))) a -> DimMat (: [*] ri (: [*] ci ([] [*]))) a -> DimMat (: [*] ri (: [*] ci ([] [*]))) a

mul :: (ZipWithZipWithMul sh sh' sh'', sh'' ~ `[_1 : _2, DOne : _3]`) => DimMat sh a -> DimMat sh' a -> DimMat sh'' a

divide :: (ZipWithZipWithMul sh' sh'' sh, sh'' ~ `[_1 : _2, DOne : _3]`) => DimMat sh a -> DimMat sh' a -> DimMat sh'' a

equal :: m ~ DimMat `[ri, ci]` a => m -> m -> Bool

arctan2 :: sh' ~ MapMapConst DOne sh => DimMat sh a -> DimMat sh a -> DimMat sh' a

hconcat :: (MultEq rem a11 b11, MapMultEq rem ri1 ri2, MapMultEq rem ci2 ci2', AppendEq ci1 ci2' ci3) => DimMat `[a11 : ri1, ci1]` a -> DimMat `[b11 : ri2, ci2]` a -> DimMat `[a11 : ri1, ci3]` a

the slightly involved type here exists because ci1 and ci2 both start with DOne, but ci2's contribution to ci3 does not need a DOne at the start. Another way to read the constraints here is:

 map (*rem) (a11 : ri) = b11 : bi
 ci3 = ci1 ++ map (*rem) ci2

The same idea happens with vconcat.

vconcat :: (AppendEq aRi (b11 : bRi') abRi, MultEq rem b11 a11, MapMultEq rem bCi aCi, MapMultEq rem bRi bRi') => DimMat `[a11 : aRi, DOne : aCi]` a -> DimMat `[b11 : bRi, DOne : bCi]` a -> DimMat `[a11 : abRi, DOne : aCi]` a

cmap :: MapMultEq deltaE ri ri' => proxy deltaE -> (forall l m t i th n j e. (NumType l, NumType m, NumType t, NumType i, NumType n, NumType j, e ~ Dim l m t i th n j) => Quantity e a -> Quantity (Mul e deltaE) b) -> DimMat `[ri, ci]` a -> DimMat `[ri', ci]` b

konst :: forall sh u us ones a _1. (Field a, NumType (Len ones), NumType (Len us), ones ~ (DOne : _1)) => Quantity u a -> DimMat `[us, ones]` a

Product class

multiply :: (Product a, sh ~ `[_1, _2]`, MultiplyCxt sh rj ri') => DimMat sh a -> DimMat (rj : cj) a -> DimMat (ri' : cj) a

does H.mXm and H.mXv.

vXm and vXv (called dot) might be supported in the future too

Random numbers

Element conversion

Input/Output

Experimental

Numeric.LinearAlgebra.Algorithms

incomplete wrapper for Numeric.LinearAlgebra.Algorithms

Linear Systems

inv :: (InvCxt sh sh', sh' ~ `[ri2 : _1, DOne : ci2]`) => DimMat sh a -> DimMat sh' a

det :: SameLengths `[ri, ci]` => DimMat `[ri, ci]` a -> Quantity (Product ri `Mul` Product ci) a

rank :: DimMat t t1 -> Int

Matrix factorizations

Singular value decomposition

Eigensystems

The Hmatrix eig factors A into P and D where A = P D inv(P) and D is diagonal.

The units for eigenvalues can be figured out:

               _____
      -1       |  c
 P D P  = A =  |r
               |

       _______
       |   d
 P   = |c
       |

       _______
       |   -1
       |  c
  -1   |   
 P   = | -1
       |d

So we can see that the dimension labeled `d-1` in P inverse is actually the same c in A. The actual units of d don't seem to matter because the `inv(d)` un-does any units that the d adds. So d can be all DOne. But another choice, such as 1/c would be more appropriate, since then you can expm your eigenvectors (not that that seems to be something people do)?

To get the row-units of A to match up, sometimes D will have units. The equation ends up as D/c = r

Please ignore the type signatures on eig eigC etc. instead look at the type of wrapEig wrapEigOnly together with the hmatrix documentation (linked).

Perhaps the convenience definitions `eig m = wrapEig H.eig m` should be in another module.

wrapEig :: (ones ~ (DOne : _1), EigCxt `[r, c]` eigVal `[cinv, ones]`, Field y, Field z) => (Matrix x -> (Vector y, Matrix z)) -> DimMat `[r, c]` x -> (DimMat `[eigVal]` y, DimMat `[cinv, ones]` z)

wrapEigOnly :: (EigCxt `[r, c]` eigVal (), Field y) => (Matrix x -> Vector y) -> DimMat `[r, c]` x -> DimMat `[eigVal]` y

type family EigCxt sh eigenValue eigenVector :: Constraint

eigenvalues and eigenvectors

eig :: (AreRecips [*] ys3 ys1, ZipWithMul * ys2 ys3 ys, Field x, SameLength' [*] [*] ys (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys, SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys1, SameLength' [*] [*] ys1 (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] ys1 ys3, SameLength' [*] [*] ys3 ys1, SameLength' [*] [*] ys3 ys2, SameLength' [*] [*] ys2 ys3, ~ * (Div DOne (Div DOne uf)) uf, ~ * (Div y (Div y (Div DOne uf))) (Div DOne uf), ~ * (Mul (Div y (Div DOne uf)) (Div DOne uf)) y) => DimMat (: [*] (: * (Div y (Div DOne uf)) ys2) (: [*] (: * (Div DOne uf) ys3) ([] [*]))) x -> (DimMat (: [*] (: * y ys) ([] [*])) (Complex Double), DimMat (: [*] (: * (Div DOne (Div DOne uf)) ys1) (: [*] (: * DOne (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2)) ([] [*]))) (Complex Double))

wrapEig H.eig

eigC :: (AreRecips [*] ys3 ys1, ZipWithMul * ys2 ys3 ys, SameLength' [*] [*] ys (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys, SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys1, SameLength' [*] [*] ys1 (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] ys1 ys3, SameLength' [*] [*] ys3 ys1, SameLength' [*] [*] ys3 ys2, SameLength' [*] [*] ys2 ys3, ~ * (Div DOne (Div DOne uf)) uf, ~ * (Div y (Div y (Div DOne uf))) (Div DOne uf), ~ * (Mul (Div y (Div DOne uf)) (Div DOne uf)) y) => DimMat (: [*] (: * (Div y (Div DOne uf)) ys2) (: [*] (: * (Div DOne uf) ys3) ([] [*]))) (Complex Double) -> (DimMat (: [*] (: * y ys) ([] [*])) (Complex Double), DimMat (: [*] (: * (Div DOne (Div DOne uf)) ys1) (: [*] (: * DOne (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2)) ([] [*]))) (Complex Double))

wrapEig H.eigC

eigH :: (AreRecips [*] ys3 ys1, ZipWithMul * ys2 ys3 ys, SameLength' [*] [*] ys (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys, SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys1, SameLength' [*] [*] ys1 (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] ys1 ys3, SameLength' [*] [*] ys3 ys1, SameLength' [*] [*] ys3 ys2, SameLength' [*] [*] ys2 ys3, ~ * (Div DOne (Div DOne uf)) uf, ~ * (Div y (Div y (Div DOne uf))) (Div DOne uf), ~ * (Mul (Div y (Div DOne uf)) (Div DOne uf)) y) => DimMat (: [*] (: * (Div y (Div DOne uf)) ys2) (: [*] (: * (Div DOne uf) ys3) ([] [*]))) (Complex Double) -> (DimMat (: [*] (: * y ys) ([] [*])) Double, DimMat (: [*] (: * (Div DOne (Div DOne uf)) ys1) (: [*] (: * DOne (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2)) ([] [*]))) (Complex Double))

wrapEig H.eigH

eigH' :: (AreRecips [*] ys3 ys1, ZipWithMul * ys2 ys3 ys, SameLength' [*] [*] ys (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys, SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys1, SameLength' [*] [*] ys1 (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] ys1 ys3, SameLength' [*] [*] ys3 ys1, SameLength' [*] [*] ys3 ys2, SameLength' [*] [*] ys2 ys3, ~ * (Div DOne (Div DOne uf)) uf, ~ * (Div y (Div y (Div DOne uf))) (Div DOne uf), ~ * (Mul (Div y (Div DOne uf)) (Div DOne uf)) y) => DimMat (: [*] (: * (Div y (Div DOne uf)) ys2) (: [*] (: * (Div DOne uf) ys3) ([] [*]))) (Complex Double) -> (DimMat (: [*] (: * y ys) ([] [*])) Double, DimMat (: [*] (: * (Div DOne (Div DOne uf)) ys1) (: [*] (: * DOne (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2)) ([] [*]))) (Complex Double))

wrapEig H.eigH'

eigR :: (AreRecips [*] ys3 ys1, ZipWithMul * ys2 ys3 ys, SameLength' [*] [*] ys (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys, SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys1, SameLength' [*] [*] ys1 (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] ys1 ys3, SameLength' [*] [*] ys3 ys1, SameLength' [*] [*] ys3 ys2, SameLength' [*] [*] ys2 ys3, ~ * (Div DOne (Div DOne uf)) uf, ~ * (Div y (Div y (Div DOne uf))) (Div DOne uf), ~ * (Mul (Div y (Div DOne uf)) (Div DOne uf)) y) => DimMat (: [*] (: * (Div y (Div DOne uf)) ys2) (: [*] (: * (Div DOne uf) ys3) ([] [*]))) Double -> (DimMat (: [*] (: * y ys) ([] [*])) (Complex Double), DimMat (: [*] (: * (Div DOne (Div DOne uf)) ys1) (: [*] (: * DOne (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2)) ([] [*]))) (Complex Double))

wrapEig H.eigR

eigS :: (AreRecips [*] ys3 ys1, ZipWithMul * ys2 ys3 ys, SameLength' [*] [*] ys (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys, SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys1, SameLength' [*] [*] ys1 (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] ys1 ys3, SameLength' [*] [*] ys3 ys1, SameLength' [*] [*] ys3 ys2, SameLength' [*] [*] ys2 ys3, ~ * (Div DOne (Div DOne uf)) uf, ~ * (Div y (Div y (Div DOne uf))) (Div DOne uf), ~ * (Mul (Div y (Div DOne uf)) (Div DOne uf)) y) => DimMat (: [*] (: * (Div y (Div DOne uf)) ys2) (: [*] (: * (Div DOne uf) ys3) ([] [*]))) Double -> (DimMat (: [*] (: * y ys) ([] [*])) Double, DimMat (: [*] (: * (Div DOne (Div DOne uf)) ys1) (: [*] (: * DOne (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2)) ([] [*]))) Double)

wrapEig H.eigS

eigS' :: (AreRecips [*] ys3 ys1, ZipWithMul * ys2 ys3 ys, SameLength' [*] [*] ys (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys, SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys1, SameLength' [*] [*] ys1 (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] ys1 ys3, SameLength' [*] [*] ys3 ys1, SameLength' [*] [*] ys3 ys2, SameLength' [*] [*] ys2 ys3, ~ * (Div DOne (Div DOne uf)) uf, ~ * (Div y (Div y (Div DOne uf))) (Div DOne uf), ~ * (Mul (Div y (Div DOne uf)) (Div DOne uf)) y) => DimMat (: [*] (: * (Div y (Div DOne uf)) ys2) (: [*] (: * (Div DOne uf) ys3) ([] [*]))) Double -> (DimMat (: [*] (: * y ys) ([] [*])) Double, DimMat (: [*] (: * (Div DOne (Div DOne uf)) ys1) (: [*] (: * DOne (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2)) ([] [*]))) Double)

wrapEig H.eigS'

eigSH :: (AreRecips [*] ys3 ys1, ZipWithMul * ys2 ys3 ys, Field z, SameLength' [*] [*] ys (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys, SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys1, SameLength' [*] [*] ys1 (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] ys1 ys3, SameLength' [*] [*] ys3 ys1, SameLength' [*] [*] ys3 ys2, SameLength' [*] [*] ys2 ys3, ~ * (Div DOne (Div DOne uf)) uf, ~ * (Div y (Div y (Div DOne uf))) (Div DOne uf), ~ * (Mul (Div y (Div DOne uf)) (Div DOne uf)) y) => DimMat (: [*] (: * (Div y (Div DOne uf)) ys2) (: [*] (: * (Div DOne uf) ys3) ([] [*]))) z -> (DimMat (: [*] (: * y ys) ([] [*])) Double, DimMat (: [*] (: * (Div DOne (Div DOne uf)) ys1) (: [*] (: * DOne (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2)) ([] [*]))) z)

wrapEig H.eigSH

eigSH' :: (AreRecips [*] ys3 ys1, ZipWithMul * ys2 ys3 ys, Field z, SameLength' [*] [*] ys (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys, SameLength' [*] [*] (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2) ys1, SameLength' [*] [*] ys1 (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2), SameLength' [*] [*] ys1 ys3, SameLength' [*] [*] ys3 ys1, SameLength' [*] [*] ys3 ys2, SameLength' [*] [*] ys2 ys3, ~ * (Div DOne (Div DOne uf)) uf, ~ * (Div y (Div y (Div DOne uf))) (Div DOne uf), ~ * (Mul (Div y (Div DOne uf)) (Div DOne uf)) y) => DimMat (: [*] (: * (Div y (Div DOne uf)) ys2) (: [*] (: * (Div DOne uf) ys3) ([] [*]))) z -> (DimMat (: [*] (: * y ys) ([] [*])) Double, DimMat (: [*] (: * (Div DOne (Div DOne uf)) ys1) (: [*] (: * DOne (MapConst * * (Dim Zero Zero Zero Zero Zero Zero Zero) ys2)) ([] [*]))) z)

wrapEig H.eigSH'

eigenvalues

eigOnlyC :: (ZipWithMul * r c eigVal, SameLength' [*] [*] eigVal c, SameLength' [*] [*] c eigVal, SameLength' [*] [*] c r, SameLength' [*] [*] r c) => DimMat (: [*] r (: [*] c ([] [*]))) (Complex Double) -> DimMat (: [*] eigVal ([] [*])) (Complex Double)

wrapEigOnly H.eigOnlyC

eigOnlyH :: (ZipWithMul * r c eigVal, SameLength' [*] [*] eigVal c, SameLength' [*] [*] c eigVal, SameLength' [*] [*] c r, SameLength' [*] [*] r c) => DimMat (: [*] r (: [*] c ([] [*]))) (Complex Double) -> DimMat (: [*] eigVal ([] [*])) Double

wrapEigOnly H.eigOnlyH

eigOnlyR :: (ZipWithMul * r c eigVal, SameLength' [*] [*] eigVal c, SameLength' [*] [*] c eigVal, SameLength' [*] [*] c r, SameLength' [*] [*] r c) => DimMat (: [*] r (: [*] c ([] [*]))) Double -> DimMat (: [*] eigVal ([] [*])) (Complex Double)

wrapEigOnly H.eigOnlyR

eigOnlyS :: (ZipWithMul * r c eigVal, SameLength' [*] [*] eigVal c, SameLength' [*] [*] c eigVal, SameLength' [*] [*] c r, SameLength' [*] [*] r c) => DimMat (: [*] r (: [*] c ([] [*]))) Double -> DimMat (: [*] eigVal ([] [*])) Double

wrapEigOnly H.eigOnlyS

eigenvalues :: (ZipWithMul * r c eigVal, Field x, SameLength' [*] [*] eigVal c, SameLength' [*] [*] c eigVal, SameLength' [*] [*] c r, SameLength' [*] [*] r c) => DimMat (: [*] r (: [*] c ([] [*]))) x -> DimMat (: [*] eigVal ([] [*])) (Complex Double)

wrapEigOnly H.eigenvalues

eigenvaluesSH :: (ZipWithMul * r c eigVal, Field x, SameLength' [*] [*] eigVal c, SameLength' [*] [*] c eigVal, SameLength' [*] [*] c r, SameLength' [*] [*] r c) => DimMat (: [*] r (: [*] c ([] [*]))) x -> DimMat (: [*] eigVal ([] [*])) Double

wrapEigOnly H.eigenvaluesSH

eigenvaluesSH' :: (ZipWithMul * r c eigVal, Field x, SameLength' [*] [*] eigVal c, SameLength' [*] [*] c eigVal, SameLength' [*] [*] c r, SameLength' [*] [*] r c) => DimMat (: [*] r (: [*] c ([] [*]))) x -> DimMat (: [*] eigVal ([] [*])) Double

wrapEigOnly H.eigenvaluesSH'

QR

Cholesky

Hessenberg

Schur

LU

Matrix functions

expm :: AreRecips ri ci => DimMat `[ri, ci]` a -> DimMat `[ri, ci]` a

Numeric.LinearAlgebra.Algorithms.expm

y t = expm (scale t a) `multiply` y0 solves the DE y' = Ay where y0 is the value of y at time 0

Nullspace

Norms

Misc

Util

Automatic Differentiation

TODO: gradients, hessians, etc.

Types for derivative towers can see hlist's HList/Data/HList/broken/Lazy.hs

Complications include the fact that AD.grad needs a traversable, but hmatrix stuff is not traversable (due needing Storable). In ipopt-hs I got around this problem by copying data. Perhaps that is the solution?

 BROKEN
 grad :: (Num a, AreRecips i iinv, H.Element a, Storable a,
           MapMultEq o iinv r) =>
         (forall s. (AD.Mode s, H.Container H.Vector (AD.AD s a),
                     Storable (AD.AD s a), H.Field (AD.AD s a))
                 => DimMat '[i] (AD.AD s a)
                 -> Quantity o (AD.AD s a))
      -> DimMat '[i] a
      -> DimMat '[r] a
 grad f (DimVec x) = DimMat (H.fromLists [AD.grad (unQty . f . DimVec . H.fromList) (H.toList x)])
     where unQty (Dimensional a) = a

diff :: Num a => (forall s. Mode s => Dimensional v x (AD s a) -> Dimensional v y (AD s a)) -> Dimensional v x a -> Dimensional v (Div x y) a

>>> let ke velocity = velocity*velocity*(1*~kilo gram)
>>> diff ke (3 *~ (metre/second))
6.0 m^-1 kg^-1 s

actually internal

toDM :: ijs -> DimMatFromTuple ijs t

just for types produced by the matD quasiquote

type family DimMatFromTuple ijs :: * -> *

given ijs :: [[Quantity a]] (except the : and [] constructors are actually (,) and (), ie. a HList), calculate a DimMat rowUnits colUnits, where the outer product of rowUnits and colUnits gives the units at each index in the ijs. The first element of colUnits is DOne.

data DimMat sh a where

Matrix with statically checked units (and dimensions). This wraps up HMatrix. The sh type parameter here contains [row,column] units. The outer product of row and column lists gives a matrix of units of the same size as the data contained within.

some pain happens to ensure that sh always follows the convention that the first element of the column units is DOne. One remaining potential error is that you can have ri ~ '[], which would make for a 0-row matrix.

Constructors

DimMat :: (Container Matrix a, Field a) => Matrix a -> DimMat `[ri, DOne : ci]` a

DimVec :: (Container Vector a, Field a) => Vector a -> DimMat `[sh]` a

type family AtEq a n b m c :: Constraint

AtEq a n b m c calculates (At a n `Mult` At b m) ~ c, but also can infer part of the a if the b and c are known

type family MapMul a xs :: [k]

type family Inner a b :: k

type family InvCxt sh sh' :: Constraint

type family SameLengths a :: Constraint

if any [k] in the list's length is known, then all other [k] lists in the outer list will be forced to have the same length

type family Product a :: k

product

type family MapRecip a :: [k]

map recip

type family ZipWithZipWithMul xs ys zs :: Constraint

zipWith (zipWith Mul) xs ys ~ zs

type family MapMapConst a xs :: [[k]]

\a xs -> map (map (const a)) xs

type family Len a :: *

length, encoded with S and Z

type family CanAddConst a m :: Constraint

class PPUnits sh

class PPUnits' sh

type family Head a :: k

type family MapDiv a xs :: [*]

\a xs -> map (/a) xs

type family AreRecips a b :: Constraint

type family ZipWithMul xs ys zs :: Constraint

xs*ys = zs: knowing two (one that is not zero) will give the third

class PairsToList a t

class DiagBlock bs t

type family MapConst a xs :: [k]

\a xs -> map (const a) xs

class SameLength' es1 es

copied from HList

type family AppendShOf a b :: *

type family MultiplyCxt sh1 sh2 sh3 :: Constraint

type MapMultEq a xs ys = (SameLengths `[xs, ys]`, MapMultEq' a xs ys)

type family Trans sh1 sh2 :: Constraint

(Trans sh1 sh2) => asserts that the two matrices are transposes

type family Tail a :: [k]